Move namespace to dedicated one and rider linting
