+++ Sakura.Framework.Tests/Graphics/Textures/ImageDecodeResidencyProbe.cs
@@ -11,6 +11,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using NUnit.Framework;
 using Sakura.Framework.Graphics.Textures;
+using Sakura.Framework.Graphics.Textures.ImageSharp;
 using Sakura.Framework.Maths;
 using SixLabors.ImageSharp;
 using SixLabors.ImageSharp.PixelFormats;
