+++ Sakura.Framework.Tests/Graphics/Textures/TextureRegistryTest.cs
